feat(max-adapters): move listeners to different files #404, #405
